// apps/web/src/app/api/model/createProjectRequest.ts
/**
 * Talos API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CreateProjectRequest {
    name: string;
    repoUrl: string;
    /**
     * Defaults to \"main\" when omitted.
     */
    defaultBranch?: string;
    stackType: string;
}